+++ scripts/macros/monsterFeatures/fireSnake/fireSnake.js
@@ -1,0 +1,4 @@
+import {heatedBody} from './heatedBody.js';
+export let fireSnake = {
+    'heatedBody': heatedBody
+}