+++ scripts/macros/monsterFeatures/salamander/salamander.js
@@ -1,0 +1,6 @@
+import {heatedBody} from '../fireSnake/heatedBody.js';
+import {tail} from './tail.js';
+export let salamander = {
+    'heatedBody': heatedBody,
+    'tail': tail
+}